+++ src/containers/sb3-loader-from-local.jsx
@@ -7,7 +7,7 @@ import {defineMessages, injectIntl, intlShape} from 'react-intl';
 import analytics from '../lib/analytics';
 import log from '../lib/log';
 import {setProjectTitle} from '../reducers/project-title';
-import {startLoadingFileUpload, doneLoadingFileUpload} from '../reducers/project-id';
+import {startLoadingFileUpload, doneLoadingFileUpload} from '../reducers/project-state';
 
 import {
     openLoadingProject,
